mrp: remove unused checkMediaRedundancy parameter
The spec says on p41, "Check Media Redundancy: This attribute specifies
whether monitoring of MRM state is enabled (TRUE) or disabled (FALSE)
in the redundancy domain.", also in p44 and p50: "Check Media Redundancy:
This parameter selects whether monitoring of MRM state is enabled or
disabled."

This is the main switch to turn on/off MRP as a whole.
We don't need it: If someone wants to turn off a MRP in a switch,
they can remove the mrp submodule from the switch via NED or ini
configuration.